% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphspme-package.R
\docType{package}
\name{StableTrees}
\alias{StableTrees}
\title{Stable Regression Tree}
\description{
\code{GraphSPME} is a package for stable updating of regression trees.
}
\details{
Important functions:

\itemize{
\item \code{\link{tree.train}}: function for training an stable regression tree model
\item \code{\link{predict.tree}}: function for predicting using the trained tree
}

See individual function documentation for usage.
}
\author{
Morten Blørstad
}